TVB-2531 Changes for compatibility with PostgreSQL

+++ framework_tvb/tvb/adapters/datatypes/db/region_mapping.py
@@ -27,6 +27,7 @@
 #   Frontiers in Neuroinformatics (7:10. doi: 10.3389/fninf.2013.00010)
 #
 #
+import numpy
 from sqlalchemy import Column, Integer, ForeignKey, Float, String
 from sqlalchemy.orm import relationship
 from tvb.core.entities.model.model_datatype import DataType, DataTypeMatrix
@@ -57,9 +58,9 @@ class RegionMappingIndex(DataType):
         # type: (RegionMapping)  -> None
         super(RegionMappingIndex, self).fill_from_has_traits(datatype)
         self.array_data_min, self.array_data_max, self.array_data_mean = from_ndarray(datatype.array_data)
-        self.array_data_min = float(self.array_data_min)
-        self.array_data_mean = float(self.array_data_mean)
-        self.array_data_max = float(self.array_data_max)
+        self.array_data_min = numpy.float64(self.array_data_min)
+        self.array_data_mean = numpy.float64(self.array_data_mean)
+        self.array_data_max = numpy.float64(self.array_data_max)
         self.surface_gid = datatype.surface.gid.hex
         self.connectivity_gid = datatype.connectivity.gid.hex
 
@@ -84,8 +85,8 @@ class RegionVolumeMappingIndex(DataTypeMatrix):
         # type: (RegionVolumeMapping)  -> None
         super(RegionVolumeMappingIndex, self).fill_from_has_traits(datatype)
         self.array_data_min, self.array_data_max, self.array_data_mean = from_ndarray(datatype.array_data)
-        self.array_data_min = float(self.array_data_min)
-        self.array_data_mean = float(self.array_data_mean)
-        self.array_data_max = float(self.array_data_max)
+        self.array_data_min = numpy.float64(self.array_data_min)
+        self.array_data_mean = numpy.float64(self.array_data_mean)
+        self.array_data_max = numpy.float64(self.array_data_max)
         self.connectivity_gid = datatype.connectivity.gid.hex
         self.volume_gid = datatype.volume.gid.hex
